+++ docs/conf.py
@@ -9,7 +9,7 @@
 project = "libunicornpkg"
 copyright = "2025, Tomodachi94"
 author = "Tomodachi94"
-release = "1.3.0"
+release = "1.3.1"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration